requirements.txt
google-cloud-dialogflow==2.1.0
python-dotenv==0.16.0
python-telegram-bot==11.1.0
vk-api==11.9.1
